Support basic blocks without a terminating jump instruction
My understanding of the QBE IL specification was that such basic
blocks are not supported, i.e. that a jump instruction is mandatory.
However, looking at the QBE test cases and at QBE code emitted by
cproc this doesn't actually seem to be the case. Meaning, the
terminating jump instruction is optional and if it is not present then
fallthrough behaviour is assumed.